Reorganize record/replay code to match dbt-adapters.
