+++ dbt/adapters/postgres/record/__init__.py
@@ -1,0 +1,2 @@
+from dbt.adapters.postgres.record.cursor.cursor import PostgresRecordReplayCursor
+from dbt.adapters.postgres.record.handle import PostgresRecordReplayHandle
